+++ cligen/clCfgInit.nim
@@ -81,6 +81,8 @@ proc apply(c: var ClCfg, path: string, plain=false) =
         of "longprefixok": c.longPfxOk = e.value.optionNormalize in yes
         of "stopprefixok": c.stopPfxOk = e.value.optionNormalize in yes
         of "argendsopts" : c.argEndsOpts = e.value.optionNormalize in yes
+        of "oneperarg"   : c.onePerArg   = e.value.optionNormalize in yes
+        of "noshort"     : c.noShort     = e.value.optionNormalize in yes
         else:
           stderr.write path&":"&" unexpected setting "&e.key&"\nExpecting: "&
             "requireseparator separatorchars longprefixok stopprefixok\n"